+++ backend/src/main/java/com/jujeol/drink/domain/RecommendForAnonymous.java
@@ -1,5 +1,6 @@
 package com.jujeol.drink.domain;
 
+import com.jujeol.drink.application.RecommendStrategy;
 import com.jujeol.drink.domain.repository.DrinkRepository;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -7,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 
 @RequiredArgsConstructor
-public class RecommendAnonymous implements RecommendState {
+public class RecommendForAnonymous implements RecommendStrategy {
 
     private final DrinkRepository drinkRepository;
