--- backend/src/main/java/com/jujeol/drink/domain/RecommendState.java
@@ -1,8 +1,0 @@
-package com.jujeol.drink.domain;
-
-import java.util.List;
-
-public interface RecommendState {
-
-    List<Drink> recommend(Long memberId, int pageSize);
-}
